Remove Auto-Download of Nightly builds
This feature was quite useful. However, this has caused a bit of an issue for AppVeyor and the DeSmuME devs as appveyor seems to now has (or always had?) a limit of 1024 MB /day download limit of artifact files.

This seems to be reached quite quickly and it could likely be that DeSmuMAR is the cause, or is at least not helping.

Therefore I have decided to just remove this feature completely. It now just tells the user to place DeSmuME next to DeSmuMAR.

This also removes the dependency on Newtonsoft.Json.